chore: made all mocks unfiltered for mysql

+++ pkg/core/proxy/integrations/mysql/command/rowscols/binaryProtocolRowPacket.go
@@ -257,9 +257,9 @@ func EncodeBinaryRow(_ context.Context, logger *zap.Logger, row *mysql.BinaryRow
 		case mysql.FieldTypeLong:
 			var val any
 			if columnEntry.Unsigned {
-				val = uint32(columnEntry.Value.(int32))
+				val = uint32(columnEntry.Value.(int))
 			} else {
-				val = columnEntry.Value.(int32)
+				val = int32(columnEntry.Value.(int))
 			}
 			if err := binary.Write(buf, binary.LittleEndian, val); err != nil {
 				return nil, fmt.Errorf("failed to write %T value: %w", val, err)
@@ -275,9 +275,9 @@ func EncodeBinaryRow(_ context.Context, logger *zap.Logger, row *mysql.BinaryRow
 		case mysql.FieldTypeTiny:
 			var val any
 			if columnEntry.Unsigned {
-				val = uint8(columnEntry.Value.(int8))
+				val = uint8(columnEntry.Value.(int))
 			} else {
-				val = columnEntry.Value.(int8)
+				val = int8(columnEntry.Value.(int))
 			}
 			if err := binary.Write(buf, binary.LittleEndian, val); err != nil {
 				return nil, fmt.Errorf("failed to write %T value: %w", val, err)
@@ -286,20 +286,19 @@ func EncodeBinaryRow(_ context.Context, logger *zap.Logger, row *mysql.BinaryRow
 		case mysql.FieldTypeShort, mysql.FieldTypeYear:
 			var val any
 			if columnEntry.Unsigned {
-				val = uint16(columnEntry.Value.(int16))
+				val = uint16(columnEntry.Value.(int))
 			} else {
-				val = columnEntry.Value.(int16)
+				val = int16(columnEntry.Value.(int))
 			}
 			if err := binary.Write(buf, binary.LittleEndian, val); err != nil {
 				return nil, fmt.Errorf("failed to write int16 value: %w", err)
 			}
 		case mysql.FieldTypeLongLong:
 			var val any
-
 			if columnEntry.Unsigned {
-				val = uint64(columnEntry.Value.(int64))
+				val = uint64(columnEntry.Value.(int))
 			} else {
-				val = columnEntry.Value.(int64)
+				val = int64(columnEntry.Value.(int))
 			}
 
 			if err := binary.Write(buf, binary.LittleEndian, val); err != nil {
